refactor: move tower sfx to effects module
